修改NE555的数据类型的BUG

+++ software/libraries/sushizhou_device/su_ne555.h
@@ -3,7 +3,7 @@
 
 #include "su_common_typedef.h"
 
-extern uint8_t frequency;//频率
+extern uint16_t frequency;//频率
 
 //必须用定时器1作为定时器，因为板子上的P34才是拿来测量频率的引脚所以计数器要用定时器0
 void Ne555TimInit(void);//NE555定时器初始化
